+++ src/lib/adapters/Errors.ts
@@ -1,8 +1,3 @@
-type Handler = (error: unknown) => void;
-
-export const handleError: Handler = (error) => {
-	/** In a real application, this would include also a call to an
-	 * error reporting service such as Sentry.
-	 **/
-	error instanceof Error ? console.error(error) : console.error(String(error));
+export const formatError = (error: unknown): string => {
+	return error instanceof Error ? error.message : String(error);
 };