+++ src/routes/photo/[id]/+page.ts
@@ -1,7 +1,7 @@
 import type { HttpError } from '@sveltejs/kit';
 import { error } from '@sveltejs/kit';
 import { HttpStatusCode } from 'axios';
-import { handleError } from '$lib/adapters/Errors';
+import { formatError } from '$lib/adapters/Errors';
 import type { Photo } from '$lib/domain/Photo';
 import { browser } from '$lib/services';
 import type { PageLoad } from './$types';
@@ -21,8 +21,8 @@ export const load: PageLoad<Response> = async ({ params }) => {
 		return {
 			photo: await browser.loadPhoto(id)
 		};
-	} catch (err: unknown) {
-		handleError(err);
+	} catch (error: unknown) {
+		console.error(formatError(error));
 		throw notFoundPhoto(id);
 	}
 };